content/base/test/test_bug429157.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=429157
-->
<head>
  <title>Test for Bug 429157</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="runtests();">

<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=429157">Mozilla Bug 429157</a>
<br>

<script>
var missingPlugins = new Array();

function pluginNotFound(event)
{
  missingPlugins.push(event.target);
}

document.addEventListener("PluginNotFound", pluginNotFound, false);
</script>

<object id="obj1" type="image/png"    >ALT image/png</object><br>
<object id="obj2" type="image/svg+xml">ALT image/svg+xml</object><br>
<object id="obj3" type="text/html"    >ALT text/html</object><br>
<object id="obj4" type="text/plain"   >ALT text/plain</object><br>

<script class="testbody" type="text/javascript">
function runtests()
{
  is(missingPlugins.length, 0, "There should be no missing plugins for this page");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
